requirements.txt
aiohttp>=3.8.0
pyyaml>=6.0
psutil>=5.9.0 
pillow==10.2.0
qrcode>=7.3.0 